backend/stats/requirements.txt
jsonschema==2.6.0
python-dotenv==0.10.1
sanic==18.12.0
asyncpg==0.18.2
